Add methods to the array instead of standalone functions.
If you want to add the computeMax as a method for the []int slice, Go does not support methods directly on slices or basic types. However, you can create a type alias for []int and then add the method to that type. In this example, we create a type IntegerArray []int which is essentially a slice of int.
The method ComputeMax() is then added to IntegerArray. Now, any IntegerArray can call the ComputeMax() method and it will return the max value and its index.